Carpool system to increase sustainability
A concept of a carpool system which was an attempt to reduce carbon emissions by giving people alternatives near them to find people near them who are going to the same place.
A single person can be either the driver or the rider depending on their choice. There was a navigation system (map) to help the users see people around them.
We looked at some edge cases like the validity of the drivers license and the overall rating for both the drivers and the rider by previous users.